post config: run depmod for alt modules for hypothetical edge case
`-alt` kernel modules for standard kernel live in `override` section,
which is activated by this package.  In the hypothetical case where
alt modules are installed before xcp-ng-release, we need to regenerate
the kmod cache.  OTOH if xcp-ng-release is configured before kmod is
installed this command could not run, and the invocation included in
the alt module rpm will do its job.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>