Move depmod execution from '%post config' to '%posttrans config'
When installing XCP-ng with kernel-alt, the kernel-alt package is
installed after xcp-ng-release-config and the call to depmod in %post
reports an error because kernel-alt modules are not installed yet.
Moving the depmod call in %posttrans ensure the modules are present when
executed.

This does not occur with main kernel because xcp-ng-release has a
dependency on kernel-livepatch package so the main kernel is installed
before xcp-ng-release.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>