WIP Provide 10.0-32.el10, pick macros.x86_64_v2 from almalinux-release-10.0-32.el10
Alma just published 10.0-32.el10, which again breaks our repo.

10.0-32.el10 changes how they install that file for x86_64_v2 build
options... and which we did not have, having started off 8.3 instead.
Previous binaries might in fact have been compiled for v3 and we did
not notice?  Possibly... or not, since that file is also supposed to
be the one "to change rpm names"...

Definitely needs digging...

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>